[Refactor] 설정 라벨 컴포넌트화
