+++ src/components/home/label/SnowColorControl.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+const SnowColorControl = ({ snowflakeColor, setSnowflakeColor }) => {
+  return (
+    <label>
+      눈송이 색깔:
+      <input
+        type="color"
+        value={snowflakeColor}
+        onChange={(e) => setSnowflakeColor(e.target.value)}
+      />
+    </label>
+  );
+};
+export default SnowColorControl;
